fix(core): repair truncated tool arguments for built-ins
