+++ codex-rs/core/src/tools/router.rs
@@ -5,6 +5,7 @@ use crate::client_common::tools::ToolSpec;
 use crate::codex::Session;
 use crate::codex::TurnContext;
 use crate::function_tool::FunctionCallError;
+use crate::tool_arguments::repair_tool_arguments;
 use crate::tools::context::SharedTurnDiffTracker;
 use crate::tools::context::ToolInvocation;
 use crate::tools::context::ToolPayload;
@@ -16,6 +17,7 @@ use codex_protocol::models::LocalShellAction;
 use codex_protocol::models::ResponseInputItem;
 use codex_protocol::models::ResponseItem;
 use codex_protocol::models::ShellToolCallParams;
+use tracing::warn;
 
 #[derive(Clone)]
 pub struct ToolCall {
@@ -76,6 +78,15 @@ impl ToolRouter {
                         },
                     }))
                 } else {
+                    let mut arguments = arguments;
+                    if let Some(fixed) = repair_tool_arguments(&arguments) {
+                        warn!(
+                            tool_name = %name,
+                            "synthesized missing closing delimiters for tool arguments"
+                        );
+                        arguments = fixed;
+                    }
+
                     let payload = if name == "unified_exec" {
                         ToolPayload::UnifiedExec { arguments }
                     } else {
